fix: schema 강화, useActionState 삭제

+++ src/app/api/proxy/k/image/schema.ts
@@ -1,7 +1,7 @@
 import { z } from 'zod/v4'
 
 export const GETProxyKImageSchema = z.object({
-  id: z.coerce.number().positive(),
+  id: z.coerce.number().int().positive(),
 })
 
 export type GETProxyKImageRequest = z.infer<typeof GETProxyKImageSchema>
